BugFix ameliorate Qt changes to binary formats for QDataStream
In other words - fix https://github.com/Mudlet/Mudlet/issues/3088 ...!

This forces the binary file format to be that which is used for Qt 5.12 if
the **run-time** version of the Qt libraries are Qt 5.13 or later. This is
needed to handle a change in the format that `QFont`s are saved/loaded in
a binary form in a `QDataStream` but it also clamps the binary format to
be equivalent to `QDataStream::Qt_5_12` everywhere I could see it used so
that future Mudlet versions do not suffer further issues going forward when
Qt revise the `QDataStream` format for the classes it handles.

This should also close https://github.com/Mudlet/Mudlet/issues/785 !

NOTE: THIS WILL BREAK THINGS TEMPORARILY FOR USERS OF MUDLET VERSIONS AFTER
4.0.1 OR THOSE WHO HAVE MANUALLY SET THE FILE FORMAT ON THEIR MAP TO BE
VERSION 19 OR HIGHER AND HAVE MOVED BETWEEN A MUDLET USING A RUN-TIME QT
VERSION LESS THAN QT 5.13 AND ONE USING THAT OR LATER. IT WILL LIKELY CAUSE
EXISTING MAP FILE CONTENTS TO BECOME GARBAGE WHEN READ BY A MUDLET
VERSION INCLUDING THIS PULL-REQUEST - SO IT IS NECESSARY TO OPEN ANY WANTED
VERSION 19 OR LATER MAP FILES IN THE CURRENT (BUGGY) QT 5.13 OR LATER USING
MUDLET AND RESAVE IT IN MUDLET FILE FORMAT 18 BEFORE UPGRADING TO A MUDLET
WITH THIS PULL-REQUEST INCLUDED. ONCE THE MAP IS THEN LOADED IN THE NEWER
MUDLET IT CAN BE RESET TO THE LATEST MAP FORMAT - and we should avoid
similar Qt library change induced problems in the future.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>