BugFixAndRevise: undo a bad change to an #include and change a constant
The bugfix element is to undo a defective change I made to the `#include`
for `qtkeychain/keychain.h`.

The constant change is to replace the literal value `18` with
`scmQDataStreamFormat_5_12` to better document what it is for in the places
it is used. It is needed because the Qt `enum` value that it represents
will not be defined in all the Qt versions we currently support.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>