Revise: place "ansi_###" 256-colours in another table
This leaves just the ANSI 16-colours in the main `color_table` but moves
the ones with a numeric suffix to another table called `ansi_color_table`
which is a table that uses an index as the key rather than a string.

The first sixteen entries in the new table (`ansi_color_table[0]` to
`ansi_color_table[15]` are updated alongside the ones in the main
color_table when they get redefined by the user or game server.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>