Enhance: hide the ansi_### colours in the output from showColors()
This was suggested during peer-review in order to hide the 256 colours
in the output but still allow them to be used in the same way as the
others.

Note that the first 16 of them are still shown in both dromedaryCase and
snake_case forms where the `color_table` key is given as `ansi` and its
default colour name.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>