Enhance: overhaul stopwatches (#2516)
This commit is a squash-and-merged of a number of commits with the
following edited summary of individual messages:

This PR allows stopwatches to be stopped and read at any time afterwards,
for any number of times. It also:
* allows stopwatches to be destroyed - so ID numbers WILL get reused - with
  a Lua (bool) destroyStopWatch((int) id) function.
* allows them to be marked as persistent so that they are saved with the
  profile and reloaded again - if they were running then they will continue
  to increment when the profile is not loaded so they can be used to
  real time events outside of the profile/session; this uses a new Lua function
  (bool) setStopWatchPersistence((int) id, (bool) setPersistent).
* allows them to be adjusted (even so they become negative) so can be used
  to count down as well as count up time - whether the stopwatch concerned
  is running or not.
* have error messages that conform to our current style
* have more run-time handling - most actions that produce no effect will
  advise that this is the case -  e.g. stopping a stopwatch that was NOT
  running...
* can handle periods of time longer than a day - the previous
  implementation wrapped around after 24 hours.
* should not affected by DST changes - OS permitting.
* the Lua API also gains a getStopWatches() function that returns a table
  with the id numbers as keys and values as tables of:
  * (bool) isRunning
  * (bool) isPersistent
  * (string) name
  * (table) elapsedTime - containing broken down time:
    * (bool) negative
    * (int) days
    * (int) hours (0 to 23)
    * (int) minutes (0 to 59)
    * (int) seconds (0 to 59)
    * (int) milliSeconds (0 to 999)
    * (float) decimalSeconds floating point value of whole of the time in
    seconds (can be negative!)
* allows stop watches to be named - which is useful as then they can be
  identified in scripts; this means that all the stop watch functions
  can now take a name string as well as a numeric argument. Using an empty
  string will access the first (lowest id) stopwatch that does not have a
  name and the createStopWatch() function will accept an optional string
  argument as a name.  For simplicity each name must be unique and this is
  enforced for that function and the added setStopWatchName(id or name,
  newName) function. The latter will also accept an empty string as either
  the first or second argument; in the first case it will assign the name
  to the first unnamed stopwatch and the second will clear the name of the
  specified one.
* added getStopWatchBrokenDownTime(...) which returns the same broken
  down elements in a table for a single specified timer (day count; hours;
  minutes; seconds; milliseconds and whether the time is positive or
  negative {when preset with a negative adjustment and used as a count
  down})...

During debugging I found out that the process of loading an existing
profile that contained createStopWatch() calls was creating them during
the testing/loading phase so I had to add extra code to prevent that
function from taking effect whilst (bool) Host::mIsProfileLoadingSequence
is true. Then, when testing the resetProfile() function I found that the
same thing was happening AND that the non-persistent stopwatches needed to
be removed as well - which is now solved by also preventing stopwatch
creation whilst (bool) Host::mResetProfile is set and by running a new
method (void) Host::removeAllNonPersistentStopWatches()!

It seems that Qt versions later than 5.7 (which was still usable when this
PR was started are happy to compare a QString with a QLatin1Char but
that version wasn't...

The code in (void) stopWatch::Host::adjustMilliSeconds(const qint64) originally
used (QDateTime) QDateTime::addMSecs(qint64) incorrectly in that I had thought
it adjusted the QDateTime it was called upon whereas it returns a reference
to the adjusted value - so needs to be invoked in a different manner.

QString stopWatch::getElapsedDayTimeString() const made use of int for
some intermediate local variables but on Windows platforms the int type may
be 32-bit long and thus not big enough to contain 64-bit values - and some
other locals can be much shorter because of the limits that the code places
on their values but will give compiler warnings without static_cast <T>.

Revise: provide backwards-compatibility for Lua startStopWatch(id) function

The prior form of startStopWatch(...) would reset and restart the indicated
stopwatch each time that it was called.  This is not really compatible with
the revised functionality which allows the recorded time to be adjusted
even before the stopwatch is first used.  To allow existing scripts to
continue to experience the same API this commit adds an optional second
boolean argument to the startStopWatch(...) call ONLY WHEN the first
argument is an id number and NOT a string name.  If the second argument is
omitted (as it will be when using older scripts) or is true then each time
the function is used the stopwatch will be reset and restarted.

If the second argument is false then the new behaviour of starting the
stopwatch only if it is not already running and continuing from whatever
the cumulative past time is on the stopwatch already is carried out - as it
will be if the first argument is a name.

Also remove some extraneous semi-colons I had accidentally put in in the
past...!

Revised changes things after peer review:

Revert to using int for the stopWatch ID number and instead use a
static_cast<int>(...) to silence the warnings about converting the long
return value of lua_tointeger(...) to an int type.

Fix a couple of typos in comments.

Remove extra detail from an error message.

Use std::chrono_literals to specify some needed time intervals rather than
large long integer literal constants.

Refactor a block of code used to generate the broken down time as a Lua
table so that two instances are handled by a single helper function.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>