BugFix: recreate prior start stop-watch when it is created behaviour
Also allow them to be created during the profile startup sequence when
lua scripts are run on loading (but after a prior compilation to test
for script validity has been done).

It wasn't obvious to me at the time but the old type of stopwatch had the
characteristic that their mere creation was sufficient to start them
running. My new design did not replicate that behaviour so it broke
existing scripts that expected this. This commit restores that action IF
`createStopWatch()` IS INVOKED WITHOUT ANY ARGUMENTS - as it would be if
existing scripts used that API call. An optional boolean argument will
prevent the auto-start if it is provided as a `false` but if a "name" for
the stopwatch to be created it is assumed that the user wants the newer
behaviour and then it will be created in a "stopped" and "reset" state.
This may be overridden by a further boolean argument which in this case
must be `true` to "autostart" the (named) stopwatch so created.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>