Modernize cmake files (part 1) (#3167)
* Modernize FindHUNSPELL

It now exports a target so we don't need to manually add variables to
the Mudlet binary and uses pkg-config enhanced search functionality
if the tool is available.

* Modernize FindPCRE

It now exports a target so we don't need to manually add variables to
the Mudlet binary and uses pkg-config enhanced search functionality
if the tool is available.

* Allow CI builds for this branch

* Modernize FindPUGIXML

* Fix and improve static finding of hunspell and PCRE libraries

* Modernize FindYAJL

* Use exported Qt targets instead of manually including variables

* Use cmake AUTO* capbilities for Qt

This makes handling the file lists nicer as we don't have to take care which
lists to put what in. And cmake is much more efficient in the way it's using the
meta-compiler things than if called manually.

* Remove manual finding of lua

This should be done by the lua find module. If that is not enough, we may have
to write our own in the future.

* Use exported target ZLIB instead of exported variables

* Some reordering and cleanup

* Modernize FindZIP.cmake

* Use exported target instead of variables

* Use own copy of FindLua51 to export targets

* Use exported target for OpenGL

* Add forgotten explicit target

* Avoid hardcoding position independent code flags

* Don't set the stdlibC++ manually

* Move most compiler flags around to modify only the mudlet target

* Undo: Allow travis builds for this branch

* Remove MSVC references in our CMakeLists.txt

* Use header-only boost target