Enhance: add support for more than 3 mouse buttons
This applies to the main console window for a profile and now reports all
mouse button presses and releases in the `sysWindowMousePressEvent` and
`sysWindowMouseReleaseEvent` instead of only the first three (left, right,
middle).

The code for the same things in Labels was checked and I found that we
were generally okay there already except that we were fabricating native
Qt mouse events for the mapper where it overlaid a label but we were using
an older (Qt4 compatible) event that only handled one mouse wheel at a time
and we were assuming it was for a vertical mouse wheel. As the owner of a
mouse with two wheels that assumption is misplaced so I have rejigged
`(bool) TLabel::forwardEventToMapper(QEvent*)` so that it can pass
horizontal and vertical mouse wheel events in the future.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>