Refactor: remove unneeded TMap::initiasliseMembersFromHost()
As the `TMap` members being initialised are (now) public they can assigned
values from the Host members they are caching directly in the
`mudlet::createMapper(...)` method. In that method - since there are many
uses of `(TMap*) pHost->mpMap` it is sensible to create a shorter named
variable `pMap` to hold that pointer value.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>