Revise: show transparency effects in custom player room marker colours
It is not possible to just colour the background of the `QPushButton`s that
set the custom colours to use for the *custom* player room marker as the
transparency effects do not show. Instead this commit arranges for an icon
with a black-and-white checker board pattern is over-painted with a brush
in the relevant colour so that when the opacity is less than 100% the
pattern is shown to an extent which is proportional to the degree of
transparency of the colour chosen.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>