Enhance: store and reproduce the editor right side splitter settings
This is done for each Mudlet item view with and without the Errors Console
being visible.

This is a proposal to reduce the impact that:
https://github.com/Mudlet/Mudlet/issues/3225 has and it also makes the size
and position for the editor be profile specific instead of being the same
for every profile which will help when using the editor and multi-playing.
Similarly the auto-save interval, being technically a per profile option is
also now stored in a per profile manner - though it is not user setable as
far as I can tell.

Also:
* swapped around an item in the `dlgTriggerEditor` constructor to fix an
 ordering error.
* renamed: `dlgTriggerEditor::slot_viewErrorsAction()`
 to: `dlgTriggerEditor::slot_toggleErrorsConsole()` to better describe its
 action.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>