BugFix: force UTF-8 coding for notepad so that all languages work
Without this, at least on Windows, a `local8Bit` encoding is used which
breaks,  for instance, on Emoji characters in the notepad. As it happens
I discovered this when working on ensuring better behaviour with using
Emoji characters in parts of Mudlet other than the main console window for
each profile.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>