Enhance: allow better control of the font use to provide (color) emojis
This is needed because forcing the use of "Noto Color Emoji" on all
versions of all OSes is just no good:

Even the latest version (September 2019) of Windows 10 cannot load it for
use generally. Also versions of Windows prior to 8.1 are (unless, it is
suggested, they have MS Office 2016 installed) incapable of displaying ANY
color emoji characters.

Also, before a particular version of font-config - I *think* it is 2.11.94
things are broken and color emojis from the Noto Color Emoji font (at
least) end up being - I discovered - being miscaled so that they were
grossly oversized. Unfortunately there are some current Linux
distributions, including Devuan 2.0 ("ASCII") that use an older version
than this.

I am not sure that FreeBSD can do color emojis but that might just be a
configuration issue on my end.

So this PR is designed to allow the font used as a substitute to provide
emojis to be selected from several specific options - and to try to select
a reasonable default for each OS/Version. As well as providing a "None"
option - which reverts Mudlet's behaviour back to that before Vadim
introduced the support for color emojis in #2835 .

As a fall-back to provide the SAME (currently only BW) Emojis for ALL cases
I have included the CC BY-SA 4.0 licenced font called:
[OpenMoji](https://www.openmoji.org/).
This includes support for Unicode 12.1 (Noto Color Emoji is only up to 11.0
even with the update I have included in this PR).

This PR deals only with fonts used in the following locations:
* Main Profile Console
* Dockable/Floating User Windows
* Sub-Console

Further PRs will be provided to address other areas where we might
reasonably expect color emoji characters to apply.  Note that this setting
has to be applied to the whole Mudlet application at a time as the Qt
facilities act on modifying the underlying application's font database - so
when you change the substitution for, say "DejaVu Sans" then whenever a
copy of that font is created for use in a widget then the changes will take
affect in that and all later uses - but it won't change any previous
instances of that font in other widgets. It also means that there are
distinct differences between passing references to `QFont`s and just
passing a *name* of a font. This is wht I have changed the name of:
`(bool) TConsole::setMiniConsoleFont(const QString&)`
to:
`(bool) TConsole::setMiniConsoleFontName(const QString&)` !

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>