Enhance: add the ability to destroy consoles without resetting a profile
It removes from the Lua API `closeWindow(...)` the need to go via an
intermediate method in the `mudlet` class.

It also takes out some instances of "return-after-else" that CLazy will
pick up on.

It will be a good idea to actually reinsert some documentation for the
previously undocumented closeUserWindow(...) to also account for the new
optional second boolean argument which has to be present and `true` for
the window/console to be deleted.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>