Enhance: improve AppStyleSheet system
This commit:
* stores each active profile's attempt to style the whole application as a
separate string
* combines all of them (and inserts some comments to separate the
contributions from each of them) before applying the composite to the
Mudlet application
* display that combination in a read-only form in a new tab in the
preferences - and also provides an editable one to allow the data for the
current profile to be revised compared to what was set by a call to the
Lua API `setAppStyleSheet(...)` function.
* allows the styleSheet set by the Lua API `setAppStyleSheet(...)` function
to be retrieved with a new `getAppStyleSheet()` call. Should an optional
`true` be provided as argument it instead returns the composite one - but
as that cannot be revised in the same manner it should be regarded as an
inspection or debugging tool (though it could allow one profile to see
whether a particular bit of styling is already present from another active
profile and avoid duplicating the same or conflicting styling element
itself!)
* applies a metaProperty of `ProfileName` to as many UI elements as I could
quickly find - which means that it ought to be possible to restrict a
profile's attempt to style the application's widgets to those belonging
only to it. This probably requires further refinement/work. As the string
that is applied to the various elements is the same as is already applied
to `TTimer`s so that they can remember which profile created them the
constant string that identifies the metaProperty has been moved/changes
from: `(static const char*) TTimer::scmProperty_HostName`
to: `(static const char*) mudlet::scmProperty_ProfileName`
so that it can be utilised more widely within the application.
* Whilst inserting the new page into the profile preference dialogue/form
I also took the liberty of rearranging a few other things/details -
including converting some `QGridLayouts` into `QVBoxLayouts` where their
was only one column or row of entries - the latter form is resilient to
rearranging its `<items>` when they get edited in the Qt Designer plugin/
stand-alone utility - which reduces the time/effort needed to remove
noise from the git repositories when that happens.

This PR also handles providing a StyleSheet from the command line which
actually results in `QApplication::styleSheet()` returning the name of a
file rather than the style-sheet content itself.

This should go a long way towards implementing the sort of arrangement I
posited in:
https://github.com/Mudlet/Mudlet/pull/3352#issuecomment-589786786

It reorders some parts of dlgProfilePreference::(en|dis)ableHostDetails()
to put some things back into the new order of the tabs in the preferences.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>