BugFix: allow boolean argument usage where createLabel(...) says it can
I tried to call:
createLabel("testLabel", 10, 10, 180, 40, true)
but was met with the error:
"createLabel: bad argument #%d type (label fillBackground as boolean/number
(0/1) expected, got boolean!)"
which is clearly not right - it turns out that the code only checked for
an actual number and not a boolean value. This error can be seen to be
repeated in the "label clickthrough" argument.

Also, the addition of an optional (user) window name as a first option
meant that various error messages were thrown out of step with the argument
number they were supposed to report.

The same error was also present in the deprecated createButton(...)
function but since we do not encourage it's use anymore it has not been
fixed here.

This PR should restructure the createLabel(...) call to be a bit more
like other functions that take an optional first argument using a local
"current" argument index variable `s` that is pre-incremented when
examining each argument in ascending order.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>