BugFix: force UTF-8 coding for notepad so that all languages work (#3303)
* BugFix: force UTF-8 coding for notepad so that all languages work

Without this, at least on Windows, a `local8Bit` encoding is used which
breaks,  for instance, on Emoji characters in the notepad. As it happens
I discovered this when working on ensuring better behaviour with using
Emoji characters in parts of Mudlet other than the main console window for
each profile.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: handle prior notes.txt files that use the local8bit encoding

It became obvious to me that the previous commit would not work where
extended ASCII characters from the used local 8 bit encoding WERE being
used in the notepad. This commit changes things to use a different file
name `"notes_utf8.txt"` to save the notes in now - and to look for that
FIRST when trying to load them...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: change a local variable name

Original was not liked by Kebap/Leris.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>