Revise: remove some uses of QPalette that might be messing up the mapper
This is for when the mapper is attached to a `TConsole`/overlays a
`TTextEdit`.

It also changes the grapheme used on the left shift/pan button in the
mapper from a U+25C0 {BLACK LEFT-POINTING TRIANGLE} to a U+25C4 {BLACK
LEFT-POINTING POINTER}, this is so it matches the U+25BA {BLACK RIGHT-
POINTING POINTER} used on the right shift/pan button. This is needed
because the "Triangle" ones are included in the Noto Color (and one or
two other) EMOJI fonts so is shown in a colorful manner in the mapper
compared to the plain black that all the other are rendered in, because
they are NOT in the Emoji fonts!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>