Refactor: make getMudletLuaDefaultPaths() return actual used details
The previous code did not link this function to the actual list of paths
used - this makes debugging when things go wrong (or when porting to a new
platform) more difficult than it need be!

This is part of my work toward migrating to MSYS2 builds on the Windows
platform.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>