Refactor: remove duplication of effort in both C++ & Lua
It turned out that as well as looping through one list of potential list
of paths to find the `luaGlobal.lua` file from the C++ core that file also
loops through another potential list of paths to find the other files that
should be in the same directory. This is pointless and was misleading as
the latter did not produce anything that the Lua API function
`getMudletLuaDefaultPaths()` would report.

Now the path (but not the file name) that is being tried from the C++ core
is also inserted as a variable `luaGlobalPath` into the Lua interpreter
where it can be used - after converting the directory separator to match
the setting for it within the Lua interpreter - to get the absolute paths
of the other files that `LuaGlobal.lua` loads.

The error reporting has been enhanced to detect more finely a failure to
read the `LuaGlobal.lua` file to say what the problem is.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>