BugFix: fix encoding used for logging.
This should close Issue #2100 .

Part of that is because we now store the text data in a `QString`
intermediate form and setting a codec on that type of `QTextStream` (as we
were but only for HTML logs) is ineffective. Instead we need to use a
codec when that data is written out to another `QTextStream` associated
with a real I/O device (file).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>