Refactor: remove unused `(QList<bool) TBuffer::dirty`
Although we maintain a flag for each line of text in the TBuffer to,
presumably indicate that it needs repainting, we never make use of it. This
suggests that it is currently redundant and can be usefully excised from
the code-base.

It seems to have been introduced around ten years ago in:
608f6d0e812d26e8c032df84d6f938cf954a555e
then revised a bit later in the commit following that one:
899491ef8ad9215f2af278cef270dee3bcd0a570
but in the five years after that (as far forward as I checked the commits
touching `TBuffer.cpp`) I did not see anything that made use of the flags
except some `qDebug()` output at a couple of points.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>