Revise: add option to control auto-login
Comment out saving and restoring login name and password either side
of package/module loading - it ***seems*** that we do not need them to
protect the details from being changed in a loaded XML file because those
pieces of information are not stored in there.

Also:
* demote dlgConnectionProfiles::slot_connectToServer() from a slot to a
normal method - as it is not connected up to the signal/slot system.
* rename some existing controls/methods that referred to "autologin" to
now use "autoLoad" as they are associated with automatic LOADing of a
profile - whereas the target of this PR *is* to do with automatic login to
the Game Server. The only issue with that is that the profile detail stored
in a file in the base directory of the profile cannot be renamed and so
the option provided in THIS PR has to be called "autoSendLogin" (c.f. the
existing "autoLogin" being used for the, now, autoLoad feature/methods).

Conflicts resolved in:
* src/ctelnet.cpp
* src/main.cpp
* src/mudlet.cpp
* src/mudlet.h

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>